Move BufferGraphExtensions to LegacyEditor project
